+++ pkg/management/postgres/webserver/remote.go
@@ -122,7 +122,7 @@ func NewRemoteWebServer(
 		server.TLSConfig = &tls.Config{
 			MinVersion: tls.VersionTLS13,
 			GetCertificate: func(_ *tls.ClientHelloInfo) (*tls.Certificate, error) {
-				return instance.ServerCertificate, nil
+				return instance.GetServerCertificate(), nil
 			},
 		}
 	}
