+++ pkg/reconciler/instance/certificate/doc.go
@@ -1,0 +1,2 @@
+// Package certificate contains the reconciler for the PostgreSQL instance manager secrets
+package certificate
